Backend/requirements.txt
flask
flask-cors
opencv-python
ultralytics
psycopg2-binary
twilio
pydub